feat: [api-handlers-decomposition-007] - Preserve events stream HTTP behavior
