+++ pkg/api/handlers_events.go
@@ -1,0 +1,79 @@
+package api
+
+import (
+	"context"
+	"errors"
+	"net/http"
+
+	factoryapi "github.com/portpowered/infinite-you/pkg/api/generated"
+	"github.com/portpowered/infinite-you/pkg/apisurface"
+	"github.com/portpowered/infinite-you/pkg/interfaces"
+	"go.uber.org/zap"
+)
+
+// GetEvents handles GET /events as a canonical factory event SSE stream.
+func (s *Server) GetEvents(w http.ResponseWriter, r *http.Request) {
+	s.getEvents(w, r, s.runtime.SubscribeFactoryEvents)
+}
+
+func (s *Server) GetEventsBySessionId(w http.ResponseWriter, r *http.Request, sessionID factoryapi.SessionID) {
+	sessionRuntime, ok := s.requireSessionRuntime(w)
+	if !ok {
+		return
+	}
+	s.getEvents(w, r, func(ctx context.Context) (*interfaces.FactoryEventStream, error) {
+		return sessionRuntime.SubscribeFactoryEventsForSession(ctx, string(sessionID))
+	})
+}
+
+func (s *Server) getEvents(
+	w http.ResponseWriter,
+	r *http.Request,
+	subscribe func(context.Context) (*interfaces.FactoryEventStream, error),
+) {
+	flusher, ok := w.(http.Flusher)
+	if !ok {
+		s.writeError(w, http.StatusInternalServerError, "streaming unsupported", "INTERNAL_ERROR")
+		return
+	}
+
+	stream, err := subscribe(r.Context())
+	if err != nil {
+		if errors.Is(err, apisurface.ErrFactorySessionNotFound) {
+			s.writeError(w, http.StatusNotFound, "factory session not found", "NOT_FOUND")
+			return
+		}
+		s.logger.Error("subscribe factory events failed", zap.Error(err))
+		s.writeError(w, http.StatusInternalServerError, "failed to subscribe to factory events", "INTERNAL_ERROR")
+		return
+	}
+
+	w.Header().Set("Content-Type", "text/event-stream")
+	w.Header().Set("Cache-Control", "no-cache")
+	w.Header().Set("Connection", "keep-alive")
+	w.Header().Set("X-Accel-Buffering", "no")
+
+	for _, event := range stream.History {
+		if err := s.writeSSEDataJSON(w, event); err != nil {
+			s.logger.Debug("write historical factory event failed", zap.Error(err))
+			return
+		}
+		flusher.Flush()
+	}
+
+	for {
+		select {
+		case <-r.Context().Done():
+			return
+		case event, ok := <-stream.Events:
+			if !ok {
+				return
+			}
+			if err := s.writeSSEDataJSON(w, event); err != nil {
+				s.logger.Debug("write live factory event failed", zap.Error(err))
+				return
+			}
+			flusher.Flush()
+		}
+	}
+}
